+++ src/main/webui/src/targets/aladin.component.css
